Simplify Reactotron config into one file
