+++ boilerplate/app/models/helpers/useStores.ts
@@ -1,5 +1,4 @@
 import { createContext, useContext, useEffect, useState } from "react"
-import { setReactotronRootStore } from "../../services/reactotron"
 import { RootStore, RootStoreModel } from "../RootStore"
 import { setupRootStore } from "./setupRootStore"
 
@@ -57,11 +56,13 @@ export const useInitialRootStore = (callback: () => void | Promise<void>) => {
     let _unsubscribe
     ;(async () => {
       // set up the RootStore (returns the state restored from AsyncStorage)
-      const { restoredState, unsubscribe } = await setupRootStore(rootStore)
+      const { unsubscribe } = await setupRootStore(rootStore)
       _unsubscribe = unsubscribe
 
       // reactotron integration with the MST root store (DEV only)
-      setReactotronRootStore(rootStore, restoredState)
+      if (__DEV__) {
+        console.tron.trackMstNode(rootStore)
+      }
 
       // let the app know we've finished rehydrating
       setRehydrated(true)
