+++ boilerplate/app/screens/DemoDebugScreen.tsx
@@ -22,17 +22,19 @@ export const DemoDebugScreen: FC<DemoTabScreenProps<"DemoDebug">> = function Dem
 
   const demoReactotron = React.useMemo(
     () => async () => {
-      console.tron.display({
-        name: "DISPLAY",
-        value: {
-          appId: Application.applicationId,
-          appName: Application.applicationName,
-          appVersion: Application.nativeApplicationVersion,
-          appBuildVersion: Application.nativeBuildVersion,
-          hermesEnabled: usingHermes,
-        },
-        important: true,
-      })
+      if (__DEV__) {
+        console.tron.display({
+          name: "DISPLAY",
+          value: {
+            appId: Application.applicationId,
+            appName: Application.applicationName,
+            appVersion: Application.nativeApplicationVersion,
+            appBuildVersion: Application.nativeBuildVersion,
+            hermesEnabled: usingHermes,
+          },
+          important: true,
+        })
+      }
     },
     [],
   )
